PID Control

+++ src/main/include/Robot.h
@@ -24,9 +24,9 @@
 class Robot : public frc::TimedRobot {
   static const int frontLeftSteerDeviceID = 1, frontLeftDriveDeviceID = 2, frontRightSteerDeviceID = 4, frontRightDriveDeviceID = 3;
   static const int rearLeftSteerDeviceID  = 5, rearLeftDriveDeviceID  = 6, rearRightSteerDeviceID  = 7, rearRightDriveDeviceID  = 8;
-  const float reductionRatio = 8.31;
+  const double reductionRatio = 8.31;
   const int WheelDiamter = 4;
-  const int WheelCircufrence = 12.566;
+  const double WheelCircufrence = 12.566;
 
   rev::CANSparkMax m_frontLeftSteerMotor {frontLeftSteerDeviceID,  rev::CANSparkMax::MotorType::kBrushless};
   rev::CANSparkMax m_frontLeftDriveMotor {frontLeftDriveDeviceID,  rev::CANSparkMax::MotorType::kBrushless};
